Rename racun package to rs.edu.raf.banka.racun
