--- racun-service/src/main/java/racun/RacunApplication.java
@@ -1,13 +1,0 @@
-package racun;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class RacunApplication {
-	
-	public static void main(String[] args) {
-		SpringApplication.run(RacunApplication.class, args);
-	}
-
-}
