--- racun-service/src/main/java/racun/enums/RacunType.java
@@ -1,6 +1,0 @@
-package racun.enums;
-
-public enum RacunType {
-
-    KES, MARGINS_RACUN;
-}
