--- racun-service/src/main/java/racun/model/Racun.java
@@ -1,27 +1,0 @@
-package racun.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import racun.enums.RacunType;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import java.util.UUID;
-
-@Entity
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Racun {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private UUID brojRacuna;
-    private RacunType tipRacuna;
-
-}
