--- racun-service/src/main/java/racun/model/Transakcija.java
@@ -1,38 +1,0 @@
-package racun.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-import java.util.Date;
-
-@Entity
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Transakcija {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "racun_id")
-    private Racun racun;
-
-    private Date datumVreme;
-
-    private long orderId;
-    private String username;
-    private String opis;
-
-    @ManyToOne
-    @JoinColumn(name = "valuta_id")
-    private Valuta valuta;
-
-    private double uplata;
-    private double isplata;
-    private double rezervisano;
-    private double rezervisanoKoristi;
-
-}
