--- racun-service/src/main/java/racun/model/Valuta.java
@@ -1,26 +1,0 @@
-package racun.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-import java.util.List;
-
-@Entity
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Valuta {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String kodValute;
-    private String nazivValute;
-    private String oznakaValute;
-    private String drzava;
-
-
-}
