--- racun-service/src/main/java/racun/repository/SredstvaKapitalRepository.java
@@ -1,16 +1,0 @@
-package racun.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import racun.model.Racun;
-import racun.model.SredstvaKapital;
-import racun.model.Valuta;
-
-
-public interface SredstvaKapitalRepository extends JpaRepository<SredstvaKapital,Long> {
-
-
-    SredstvaKapital findByRacunAndValuta(Racun racun, Valuta valuta);
-
-
-
-}
