--- racun-service/src/main/java/racun/repository/ValutaRepository.java
@@ -1,9 +1,0 @@
-package racun.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import racun.model.Valuta;
-
-public interface ValutaRepository extends JpaRepository<Valuta,Long> {
-
-    Valuta findValutaByOznakaValute(String oznaka);
-}
