--- racun-service/src/main/java/racun/requests/TransakcijaRequest.java
@@ -1,16 +1,0 @@
-package racun.requests;
-
-import lombok.Data;
-
-import java.util.UUID;
-
-@Data
-public class TransakcijaRequest {
-
-    private UUID brojRacuna;
-    private String opis;
-    private String valutaOznaka;
-    private double iznos;
-
-
-}
