+++ racun-service/src/main/java/rs/edu/raf/banka/racun/RacunApplication.java
@@ -1,0 +1,13 @@
+package rs.edu.raf.banka.racun;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RacunApplication {
+	
+	public static void main(String[] args) {
+		SpringApplication.run(RacunApplication.class, args);
+	}
+
+}
