+++ racun-service/src/main/java/rs/edu/raf/banka/racun/enums/RacunType.java
@@ -1,0 +1,6 @@
+package rs.edu.raf.banka.racun.enums;
+
+public enum RacunType {
+
+    KES, MARGINS_RACUN;
+}
