+++ racun-service/src/main/java/rs/edu/raf/banka/racun/model/Racun.java
@@ -1,0 +1,27 @@
+package rs.edu.raf.banka.racun.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import rs.edu.raf.banka.racun.enums.RacunType;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import java.util.UUID;
+
+@Entity
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Racun {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private UUID brojRacuna;
+    private RacunType tipRacuna;
+
+}
