+++ racun-service/src/main/java/rs/edu/raf/banka/racun/repository/SredstvaKapitalRepository.java
@@ -1,0 +1,16 @@
+package rs.edu.raf.banka.racun.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import rs.edu.raf.banka.racun.model.Racun;
+import rs.edu.raf.banka.racun.model.SredstvaKapital;
+import rs.edu.raf.banka.racun.model.Valuta;
+
+
+public interface SredstvaKapitalRepository extends JpaRepository<SredstvaKapital,Long> {
+
+
+    SredstvaKapital findByRacunAndValuta(Racun racun, Valuta valuta);
+
+
+
+}
