+++ racun-service/src/main/java/rs/edu/raf/banka/racun/repository/ValutaRepository.java
@@ -1,0 +1,9 @@
+package rs.edu.raf.banka.racun.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import rs.edu.raf.banka.racun.model.Valuta;
+
+public interface ValutaRepository extends JpaRepository<Valuta,Long> {
+
+    Valuta findValutaByOznakaValute(String oznaka);
+}
