+++ racun-service/src/main/java/rs/edu/raf/banka/racun/requests/RezervacijaRequest.java
@@ -1,0 +1,14 @@
+package rs.edu.raf.banka.racun.requests;
+
+import lombok.Data;
+
+@Data
+public class RezervacijaRequest {
+
+    private String racun;
+    private String valuta;
+    private String opis;
+    private float isplata;
+    private float RezervacijaKoristi;
+    private long hartijeOdVrednostiID;
+}
