+++ racun-service/src/main/java/rs/edu/raf/banka/racun/requests/TransakcijaRequest.java
@@ -1,0 +1,16 @@
+package rs.edu.raf.banka.racun.requests;
+
+import lombok.Data;
+
+import java.util.UUID;
+
+@Data
+public class TransakcijaRequest {
+
+    private UUID brojRacuna;
+    private String opis;
+    private String valutaOznaka;
+    private double iznos;
+
+
+}
