+++ racun-service/src/main/java/rs/edu/raf/banka/racun/service/impl/RacunService.java
@@ -1,0 +1,36 @@
+package rs.edu.raf.banka.racun.service.impl;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import rs.edu.raf.banka.racun.enums.RacunType;
+import rs.edu.raf.banka.racun.model.Racun;
+import rs.edu.raf.banka.racun.repository.RacunRepository;
+
+import java.util.UUID;
+
+
+@Service
+public class RacunService {
+
+    private RacunRepository racunRepository;
+    private final SredstvaKapitalService sredstvaKapitalService;
+
+    @Autowired
+    public RacunService(RacunRepository racunRepository, SredstvaKapitalService sredstvaKapitalService){
+        this.racunRepository = racunRepository;
+        this.sredstvaKapitalService = sredstvaKapitalService;
+    }
+
+    public Racun createRacun(){
+        Racun racun = new Racun();
+        racun.setBrojRacuna(UUID.randomUUID());
+        racun.setTipRacuna(RacunType.KES);
+        racunRepository.save(racun);
+        sredstvaKapitalService.updateStanje(racun.getBrojRacuna(),1000,0,0,"RSD",0); //Pocetno stanje za testiranje
+
+        return racun;
+    }
+
+
+
+}
