+++ racun-service/src/main/java/rs/edu/raf/banka/racun/utils/DateUtils.java
@@ -1,0 +1,19 @@
+package rs.edu.raf.banka.racun.utils;
+
+import java.time.ZonedDateTime;
+import java.util.Date;
+
+public class DateUtils {
+
+    public static boolean isDateInDecayDays(Date date, Integer days) {
+        ZonedDateTime now = ZonedDateTime.now();
+        ZonedDateTime daysAgo = now.plusDays(-1L * days);
+
+        return date.toInstant().isBefore(daysAgo.toInstant());
+    }
+
+    private DateUtils() {
+
+    }
+
+}
